[gw] cleanup hubbard dimer tests
